requirements.txt
Django==4.2.10
Pillow==10.2.0
requests==2.31.0
python-dotenv==1.0.1
ebaysdk==2.2.0
numpy==1.26.4
opencv-python==4.9.0.80
scikit-image==0.22.0 